requirements.txt
bcrypt==3.1.7
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==20.0.4
httplib2==0.17.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.12
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==0.16.0
Whoosh==2.7.4
WTForms==2.2.1
